Move final files
