+++ contribs/gnogenesis/internal/validator/validator_add.go
@@ -1,4 +1,4 @@
-package main
+package validator
 
 import (
 	"context"
