+++ contribs/gnogenesis/internal/validator/validator_remove_test.go
@@ -1,10 +1,10 @@
-package main
+package validator
 
 import (
 	"context"
 	"testing"
 
-	"github.com/gnolang/contribs/gnogenesis/internal/balances"
+	"github.com/gnolang/contribs/gnogenesis/internal/common"
 	"github.com/gnolang/gno/tm2/pkg/bft/types"
 	"github.com/gnolang/gno/tm2/pkg/commands"
 	"github.com/gnolang/gno/tm2/pkg/testutils"
@@ -19,9 +19,8 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		t.Parallel()
 
 		// Create the command
-		cmd := newGenesisCmd(commands.NewTestIO())
+		cmd := NewValidatorCmd(commands.NewTestIO())
 		args := []string{
-			"validator",
 			"remove",
 			"--genesis-path",
 			"dummy-path",
@@ -29,7 +28,7 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 
 		// Run the command
 		cmdErr := cmd.ParseAndRun(context.Background(), args)
-		assert.ErrorContains(t, cmdErr, balances.errUnableToLoadGenesis.Error())
+		assert.ErrorContains(t, cmdErr, common.ErrUnableToLoadGenesis.Error())
 	})
 
 	t.Run("invalid validator address", func(t *testing.T) {
@@ -38,13 +37,12 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		tempGenesis, cleanup := testutils.NewTestFile(t)
 		t.Cleanup(cleanup)
 
-		genesis := GetDefaultGenesis()
+		genesis := common.GetDefaultGenesis()
 		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
 
 		// Create the command
-		cmd := newGenesisCmd(commands.NewTestIO())
+		cmd := NewValidatorCmd(commands.NewTestIO())
 		args := []string{
-			"validator",
 			"remove",
 			"--genesis-path",
 			tempGenesis.Name(),
@@ -63,8 +61,8 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		tempGenesis, cleanup := testutils.NewTestFile(t)
 		t.Cleanup(cleanup)
 
-		dummyKeys := GetDummyKeys(t, 2)
-		genesis := GetDefaultGenesis()
+		dummyKeys := common.GetDummyKeys(t, 2)
+		genesis := common.GetDefaultGenesis()
 
 		// Set an existing validator
 		genesis.Validators = append(genesis.Validators, types.GenesisValidator{
@@ -77,9 +75,8 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
 
 		// Create the command
-		cmd := newGenesisCmd(commands.NewTestIO())
+		cmd := NewValidatorCmd(commands.NewTestIO())
 		args := []string{
-			"validator",
 			"remove",
 			"--genesis-path",
 			tempGenesis.Name(),
@@ -98,9 +95,9 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		tempGenesis, cleanup := testutils.NewTestFile(t)
 		t.Cleanup(cleanup)
 
-		dummyKey := GetDummyKey(t)
+		dummyKey := common.GetDummyKey(t)
 
-		genesis := GetDefaultGenesis()
+		genesis := common.GetDefaultGenesis()
 
 		// Set an existing validator
 		genesis.Validators = append(genesis.Validators, types.GenesisValidator{
@@ -113,9 +110,8 @@ func TestGenesis_Validator_Remove(t *testing.T) {
 		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
 
 		// Create the command
-		cmd := newGenesisCmd(commands.NewTestIO())
+		cmd := NewValidatorCmd(commands.NewTestIO())
 		args := []string{
-			"validator",
 			"remove",
 			"--genesis-path",
 			tempGenesis.Name(),
