--- contribs/gnogenesis/types.go
@@ -1,2 +1,0 @@
-package main
-
